chore: 주석 다는 작업 추가 [#256]
